+++ editor/file_system/editor_file_system.cpp
@@ -3447,10 +3447,10 @@ bool EditorFileSystem::_should_skip_directory(const String &p_path) {
 		return true;
 	}
 
-	if (FileAccess::exists(p_path.path_join("project.godot"))) {
+	if (FileAccess::exists(p_path.path_join("project.tekisasu"))) {
 		// Skip if another project inside this.
 		if (EditorFileSystem::get_singleton() == nullptr || EditorFileSystem::get_singleton()->first_scan) {
-			WARN_PRINT_ONCE(vformat("Detected another project.godot at %s. The folder will be ignored.", p_path));
+			WARN_PRINT_ONCE(vformat("Detected another project.tekisasu at %s. The folder will be ignored.", p_path));
 		}
 		return true;
 	}